requirements.txt
numpy==1.12
scipy==0.19.1
scikit-learn==0.18.1
matplotlib
opencv==3.1
jupyter==1.0
notebook>=5.7.2
pandas==0.22
theano
keras==2.1.5
mkl-service==1.1.2
python3-pip
opencv-python==4.1.1.26
hcsr04sensor
cachetools==3.1.1 
futures==3.3.0 
google-api-core==1.14.2 
google-auth==1.6.3 
google-cloud-vision==0.39.0 
googleapis-common-protos==1.6.0 
grpcio==1.23.0 
protobuf==3.9.1 
pyasn1==0.4.7 
pyasn1-modules==0.2.6 
rsa==4.0
rpi-gpio==0.6.5
python3-w1thermsensor
Bluetin-Echo==0.2.0
python3-smbus
